inline ProfileScreen styles and remove unused Spinner export
