+++ app/(tabs)/profile.tsx
@@ -1,29 +1,40 @@
 import { useRouter } from "expo-router";
 import React, { PropsWithChildren } from "react";
-import { ImageBackground, RefreshControl, View, Image } from "react-native";
+import {
+  ImageBackground,
+  RefreshControl,
+  View,
+  Image,
+  ScrollView,
+  Text,
+  FlatList,
+  StyleSheet,
+} from "react-native";
 
 import { black50, white } from "yep/colors";
 import { Header } from "yep/components/Header";
 import { PosterAndTitle } from "yep/components/PosterAndTitle";
 import { PressableOpacity } from "yep/components/PressableOpacity";
 import { useGetViewerQuery } from "yep/graphql/generated";
-import {
-  OuterContainer,
-  InnerContainer,
-  UserInfoAndStatsContainer,
-  Avatar,
-  UserInfoRow,
-  Username,
-  StatsRow,
-  Stat,
-  ListHeader,
-  makeListWithType,
-  FavoriteContainer,
-  EverythingButTheCTA,
-} from "yep/screens/ProfileScreen/styles";
 import { StringCase, getString } from "yep/strings";
 import { darkTheme } from "yep/themes";
+import { Manrope } from "yep/typefaces";
 import { notEmpty, getTitle } from "yep/utils";
+
+type StatProps = { label: string; value: number };
+
+function Stat({ label, value }: StatProps) {
+  return (
+    <View style={styles.statContainer}>
+      <Text style={styles.statLabel} numberOfLines={1}>
+        {label}
+      </Text>
+      <Text style={styles.statValue} numberOfLines={1}>
+        {value}
+      </Text>
+    </View>
+  );
+}
 
 export default function Profile() {
   const router = useRouter();
@@ -33,7 +44,7 @@ export default function Profile() {
     refetch,
   } = useGetViewerQuery({ notifyOnNetworkStatusChange: true });
   const animeList = (viewerData?.Viewer?.favourites?.anime?.nodes ?? []).filter(
-    notEmpty
+    notEmpty,
   );
   const characterList = (
     viewerData?.Viewer?.favourites?.characters?.nodes ?? []
@@ -41,9 +52,6 @@ export default function Profile() {
 
   type AnimeItem = (typeof animeList)[number];
   type CharacterItem = (typeof characterList)[number];
-
-  const FavoriteAnimeList = makeListWithType<AnimeItem>();
-  const FavoriteCharacterList = makeListWithType<CharacterItem>();
 
   function OptionalBackgroundImage({ children }: PropsWithChildren) {
     return viewerData?.Viewer?.bannerImage ? (
@@ -59,7 +67,9 @@ export default function Profile() {
   }
 
   return (
-    <OuterContainer style={{ backgroundColor: darkTheme.background }}>
+    <View
+      style={[styles.outerContainer, { backgroundColor: darkTheme.background }]}
+    >
       <Header
         label={getString("profile", StringCase.TITLE)}
         rightSlot={
@@ -75,7 +85,8 @@ export default function Profile() {
           </PressableOpacity>
         }
       />
-      <InnerContainer
+      <ScrollView
+        contentContainerStyle={styles.innerContainerContent}
         refreshControl={
           <RefreshControl
             refreshing={loadingViewer}
@@ -87,17 +98,21 @@ export default function Profile() {
         showsVerticalScrollIndicator={false}
       >
         {viewerData?.Viewer ? (
-          <EverythingButTheCTA>
+          <View style={styles.everythingButTheCTA}>
             <OptionalBackgroundImage>
-              <UserInfoAndStatsContainer
-                style={{
-                  backgroundColor: viewerData?.Viewer?.bannerImage
-                    ? black50
-                    : darkTheme.listItemBackground,
-                }}
+              <View
+                style={[
+                  styles.userInfoAndStatsContainer,
+                  {
+                    backgroundColor: viewerData?.Viewer?.bannerImage
+                      ? black50
+                      : darkTheme.listItemBackground,
+                  },
+                ]}
               >
-                <UserInfoRow>
-                  <Avatar
+                <View style={styles.userInfoRow}>
+                  <Image
+                    style={styles.avatar}
                     source={
                       viewerData?.Viewer?.avatar
                         ? {
@@ -108,11 +123,11 @@ export default function Profile() {
                         : require("yep/assets/icons/avatar-placeholder.png")
                     }
                   />
-                  <Username numberOfLines={1}>
+                  <Text style={styles.username} numberOfLines={1}>
                     {viewerData.Viewer.name}
-                  </Username>
-                </UserInfoRow>
-                <StatsRow>
+                  </Text>
+                </View>
+                <View style={styles.statsRow}>
                   <Stat
                     label="Total Anime"
                     value={viewerData.Viewer.statistics?.anime?.count ?? 0}
@@ -123,23 +138,26 @@ export default function Profile() {
                       (viewerData.Viewer.statistics?.anime?.minutesWatched ??
                         0) /
                         60 /
-                        24
+                        24,
                     )}
                   />
-                </StatsRow>
-              </UserInfoAndStatsContainer>
+                </View>
+              </View>
             </OptionalBackgroundImage>
             {animeList.length ? (
               <View>
-                <ListHeader>Favorite Anime</ListHeader>
-                <FavoriteAnimeList
-                  contentContainerStyle={{ gap: 8 }}
+                <Text style={styles.listHeader}>Favorite Anime</Text>
+                <FlatList
+                  contentContainerStyle={[
+                    styles.listContentContainer,
+                    { gap: 8 },
+                  ]}
                   horizontal
                   showsHorizontalScrollIndicator={false}
                   keyExtractor={(item: AnimeItem) => `${item.id}`}
                   data={animeList}
                   renderItem={({ item }: { item: AnimeItem }) => (
-                    <FavoriteContainer>
+                    <View style={styles.favoriteContainer}>
                       <PressableOpacity
                         onPress={() => router.push(`/details/${item.id}`)}
                       >
@@ -149,22 +167,25 @@ export default function Profile() {
                           title={getTitle(item.title)}
                         />
                       </PressableOpacity>
-                    </FavoriteContainer>
+                    </View>
                   )}
                 />
               </View>
             ) : null}
             {characterList.length ? (
               <View>
-                <ListHeader>Favorite Characters</ListHeader>
-                <FavoriteCharacterList
+                <Text style={styles.listHeader}>Favorite Characters</Text>
+                <FlatList
                   horizontal
                   showsHorizontalScrollIndicator={false}
-                  contentContainerStyle={{ gap: 8 }}
+                  contentContainerStyle={[
+                    styles.listContentContainer,
+                    { gap: 8 },
+                  ]}
                   keyExtractor={(item: CharacterItem) => `${item.id}`}
                   data={characterList}
                   renderItem={({ item }: { item: CharacterItem }) => (
-                    <FavoriteContainer>
+                    <View style={styles.favoriteContainer}>
                       <PressableOpacity
                         onPress={() => router.push(`/character/${item.id}`)}
                       >
@@ -174,14 +195,80 @@ export default function Profile() {
                           title={item.name?.full ?? undefined}
                         />
                       </PressableOpacity>
-                    </FavoriteContainer>
+                    </View>
                   )}
                 />
               </View>
             ) : null}
-          </EverythingButTheCTA>
+          </View>
         ) : null}
-      </InnerContainer>
-    </OuterContainer>
+      </ScrollView>
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  outerContainer: {
+    flex: 1,
+  },
+  innerContainerContent: {
+    padding: 16,
+    justifyContent: "space-between",
+    gap: 16,
+  },
+  userInfoAndStatsContainer: {
+    padding: 16,
+    borderRadius: 8,
+    gap: 16,
+    overflow: "hidden",
+    flex: 1,
+  },
+  avatar: {
+    height: 40,
+    width: 40,
+    borderRadius: 20,
+  },
+  userInfoRow: {
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 8,
+  },
+  username: {
+    fontFamily: Manrope.semiBold,
+    fontSize: 20,
+    color: darkTheme.text,
+  },
+  statsRow: {
+    flexDirection: "row",
+  },
+  statContainer: {
+    flex: 1,
+    gap: 4,
+  },
+  statLabel: {
+    fontFamily: Manrope.regular,
+    fontSize: 12.8,
+    color: darkTheme.text,
+  },
+  statValue: {
+    fontFamily: Manrope.semiBold,
+    fontSize: 20,
+    color: darkTheme.text,
+  },
+  listHeader: {
+    fontFamily: Manrope.semiBold,
+    fontSize: 25,
+    color: darkTheme.text,
+    marginBottom: 8,
+  },
+  listContentContainer: {
+    alignItems: "flex-start",
+  },
+  favoriteContainer: {
+    alignItems: "center",
+  },
+  everythingButTheCTA: {
+    flex: 1,
+    gap: 16,
+  },
+});
